+++ CHANGES.md
@@ -22,6 +22,7 @@ Version 1.17 [????.??.??]
   patterns. That this function ever did extract type variables was a mistake,
   and the new behavior of `extractBoundNamesDPat` brings it in line with the
   behavior `extractBoundNamesPat`.
+* Add `Foldable` and `Traversable` instances for `DTyVarBndrSpec`.
 
 Version 1.16 [2023.10.13]
 -------------------------